feat: código quebrado
