+++ src/main/kotlin/br/com/liviafort/deliverysystem/application/resources/ReusableComponents.kt
@@ -1,67 +1,67 @@
 package br.com.liviafort.deliverysystem.application.resources
 
-import br.com.liviafort.deliverysystem.di.DependencyContainer
-import br.com.liviafort.deliverysystem.domain.customer.Customer
-import br.com.liviafort.deliverysystem.domain.customer.CustomerServiceImpl
-import br.com.liviafort.deliverysystem.domain.deliveryman.Deliveryman
-import br.com.liviafort.deliverysystem.domain.deliveryman.DeliverymanServiceImpl
-import br.com.liviafort.deliverysystem.domain.restaurant.Restaurant
-import br.com.liviafort.deliverysystem.domain.restaurant.RestaurantItem
-import br.com.liviafort.deliverysystem.domain.restaurant.RestaurantServiceImpl
-
-
-class ReusableComponents() {
-    private val customerService = CustomerServiceImpl(DependencyContainer.customerRepository)
-    private val restaurantService = RestaurantServiceImpl(DependencyContainer.restaurantRepository)
-    private val deliverymanService = DeliverymanServiceImpl(DependencyContainer.deliverymanRepository)
-
-    init {
-        createRestaurant()
-        createCustomer()
-        createDeliveryman()
-    }
-
-    private fun createRestaurant() {
-        val items1 = mutableSetOf(
-            RestaurantItem(name = "Pizza Quatro Queijos", price = 49.60),
-            RestaurantItem(name = "Pizza Diavola", price = 47.20),
-        )
-        val items2 = mutableSetOf(RestaurantItem(name = "Double cheese", price = 49.60))
-
-        val restaurant1 = Restaurant(
-            name = "Pizzaria Straj",
-            address = "Rua Alemida Junior, 45",
-            category = "Pizzaria",
-            cnpj = "38293/0003-123",
-            items = items1
-        )
-        val restaurant2 = Restaurant(
-            name = "Hamburgueria Aureau",
-            address = "Rua Caimbra, 121",
-            category = "Hamburgueria",
-            cnpj = "33493/086786-123",
-            items = items2
-        )
-
-        restaurantService.create(restaurant1)
-        restaurantService.create(restaurant2)
-    }
-
-    private fun createCustomer() {
-        val customer1 = Customer(name = "livia", phone = "12345678", address = "Miguel Seabra, 324")
-        val customer2 = Customer(name = "maria", phone = "921749124", address = "Aluisio Mendes, 324")
-
-        customerService.create(customer1)
-        customerService.create(customer2)
-
-    }
-
-    private fun createDeliveryman() {
-        val deliveryman1 = Deliveryman(name = "Jeremias", phone = "829134342", vehicle = "Car")
-        val deliveryman2 = Deliveryman(name = "Rivaldo", phone = "238423243", vehicle = "Motocicleta")
-
-        deliverymanService.create(deliveryman1)
-        deliverymanService.create(deliveryman2)
-    }
-}
+//import br.com.liviafort.deliverysystem.di.DependencyContainer
+//import br.com.liviafort.deliverysystem.domain.customer.Customer
+//import br.com.liviafort.deliverysystem.domain.customer.CustomerServiceImpl
+//import br.com.liviafort.deliverysystem.domain.deliveryman.Deliveryman
+//import br.com.liviafort.deliverysystem.domain.deliveryman.DeliverymanServiceImpl
+//import br.com.liviafort.deliverysystem.domain.restaurant.Restaurant
+//import br.com.liviafort.deliverysystem.domain.restaurant.RestaurantItem
+//import br.com.liviafort.deliverysystem.domain.restaurant.RestaurantServiceImpl
+//
+//
+//class ReusableComponents() {
+//    private val customerService = CustomerServiceImpl(DependencyContainer.customerRepository)
+//    private val restaurantService = RestaurantServiceImpl(DependencyContainer.restaurantRepository)
+//    private val deliverymanService = DeliverymanServiceImpl(DependencyContainer.deliverymanRepository)
+//
+//    init {
+//        createRestaurant()
+//        createCustomer()
+//        createDeliveryman()
+//    }
+//
+//    private fun createRestaurant() {
+//        val items1 = mutableSetOf(
+//            RestaurantItem(name = "Pizza Quatro Queijos", price = 49.60),
+//            RestaurantItem(name = "Pizza Diavola", price = 47.20),
+//        )
+//        val items2 = mutableSetOf(RestaurantItem(name = "Double cheese", price = 49.60))
+//
+//        val restaurant1 = Restaurant(
+//            name = "Pizzaria Straj",
+//            address = "Rua Alemida Junior, 45",
+//            category = "Pizzaria",
+//            cnpj = "38293/0003-123",
+//            items = items1
+//        )
+//        val restaurant2 = Restaurant(
+//            name = "Hamburgueria Aureau",
+//            address = "Rua Caimbra, 121",
+//            category = "Hamburgueria",
+//            cnpj = "33493/086786-123",
+//            items = items2
+//        )
+//
+//        restaurantService.create(restaurant1)
+//        restaurantService.create(restaurant2)
+//    }
+//
+//    private fun createCustomer() {
+//        val customer1 = Customer(name = "livia", phone = "12345678", address = "Miguel Seabra, 324")
+//        val customer2 = Customer(name = "maria", phone = "921749124", address = "Aluisio Mendes, 324")
+//
+//        customerService.create(customer1)
+//        customerService.create(customer2)
+//
+//    }
+//
+//    private fun createDeliveryman() {
+//        val deliveryman1 = Deliveryman(name = "Jeremias", phone = "829134342", vehicle = "Car")
+//        val deliveryman2 = Deliveryman(name = "Rivaldo", phone = "238423243", vehicle = "Motocicleta")
+//
+//        deliverymanService.create(deliveryman1)
+//        deliverymanService.create(deliveryman2)
+//    }
+//}
