--- src/main/kotlin/br/com/liviafort/deliverysystem/application/resources/Singleton.kt
@@ -1,21 +1,0 @@
-package br.com.liviafort.deliverysystem.application.resources
-
-//object OrderServiceSingleton {
-//    val instance = OrderServiceImpl(OrderRepositoryInMemory())
-//}
-//
-//object RestaurantServiceSingleton {
-//    val instance = RestaurantServiceImpl(RestaurantRepositoryInMemory())
-//}
-//
-//object DeliverymanServiceSingleton {
-//    val instance = DeliverymanServiceImpl(DeliverymanRepositoryInMemory())
-//}
-//
-//object CustomerServiceSingleton {
-//    val instance = CustomerServiceImpl(CustomerRepositoryInMemory())
-//}
-//
-//object RouteServiceSingleton {
-//    val instance = RouteServiceImpl(RouteRepositoryInMemory())
-//}
